Update README.md to correct the way we are checking browser for compatibility with Worker
In question 49, w is being declared without a proper keyword, which is now fixed by declaring the same with 'let' keyword

We also are checking if w is undefined, which is not necessary since it will be reinitialized. The missing component here is a check for browser compatibility with web workers, which is added now.